Rename module PublicKey to KeyPair
To add classes KeyPair, SecretKey, and Seed to this module.
