+++ src/index.ts
@@ -17,7 +17,7 @@ export { checksum, validate } from './modules/utils/CRC16';
 export { SodiumHelper } from './modules/utils/SodiumHelper';
 export { Hash, hash, hashMulti, makeUTXOKey } from './modules/data/Hash';
 export { PreImage  } from './modules/data/PreImage';
-export { PublicKey } from './modules/data/PublicKey';
+export { PublicKey } from './modules/data/KeyPair';
 export { Signature } from './modules/data/Signature';
 export { Validator } from './modules/data/Validator';
